Refactored common code into DocumentPostPageType shared between question and post
